% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{aggregate2_Duration_meansd}
\alias{aggregate2_Duration_meansd}
\title{Compute Mean and SD of Duration}
\usage{
aggregate2_Duration_meansd(df, ...)
}
\arguments{
\item{df}{The data frame with an evaluated \code{dur} numeric column.}

\item{...}{Additional arguments.}
}
\value{
A 1-row data frame with \code{mean_dur} and \code{sd_dur}.
}
\description{
Computes the mean and standard deviation of durations across all valid intervals. Automatically initializes a \code{.valid} flag if none exists.
}
