% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{aggregate2_NNsumDurEnmoEnmomaxStepStepmax}
\alias{aggregate2_NNsumDurEnmoEnmomaxStepStepmax}
\title{Title}
\usage{
aggregate2_NNsumDurEnmoEnmomaxStepStepmax(df, ...)
}
\arguments{
\item{df}{}

\item{...}{}
}
\description{
Title
}
