% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{aggregate_Duration_meanByHour}
\alias{aggregate_Duration_meanByHour}
\title{Title}
\usage{
aggregate_Duration_meanByHour(df, wide = FALSE, ...)
}
\arguments{
\item{df}{}

\item{wide}{}

\item{...}{}
}
\description{
Title
}
