% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{aggregate_DurationRel}
\alias{aggregate_DurationRel}
\title{Title}
\usage{
aggregate_DurationRel(dfCruce, dfName, units = "secs", .by = "name", ...)
}
\arguments{
\item{dfCruce}{}

\item{dfName}{}

\item{units}{}

\item{.by}{}

\item{...}{}
}
\description{
Title
}
