% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{aggregate_DurStepCadENMO}
\alias{aggregate_DurStepCadENMO}
\title{Title}
\usage{
aggregate_DurStepCadENMO(df, driver, units = "secs", ...)
}
\arguments{
\item{df}{}

\item{driver}{}

\item{units}{}

\item{...}{}
}
\value{
name, steps, dur, cad,  ENMO
}
\description{
Title
}
