% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{aggregate_ENMOCountStepCad}
\alias{aggregate_ENMOCountStepCad}
\title{Title}
\usage{
aggregate_ENMOCountStepCad(df, driver, units = "secs", ...)
}
\arguments{
\item{df}{}

\item{driver}{}

\item{units}{}

\item{...}{}
}
\description{
Title
}
