% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{aggregate_NDurEnmoStep}
\alias{aggregate_NDurEnmoStep}
\title{Title}
\usage{
aggregate_NDurEnmoStep(df, dfDriver, units = "mins", ...)
}
\arguments{
\item{df}{}

\item{dfDriver}{}

\item{units}{}

\item{...}{}
}
\description{
Title
}
