% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{aggregate_Quantile}
\alias{aggregate_Quantile}
\title{Title}
\usage{
aggregate_Quantile(df, probs = c(0, 0.25, 0.5, 0.75, 1), .by = "name", ...)
}
\arguments{
\item{df}{}

\item{probs}{}

\item{.by}{}

\item{...}{}
}
\description{
Title
}
