% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{aggregateVariablesLong}
\alias{aggregateVariablesLong}
\title{Title}
\usage{
aggregateVariablesLong(intervals, variablesLong, driver, ...)
}
\arguments{
\item{intervals}{}

\item{variablesLong}{}

\item{driver}{}

\item{...}{}
}
\description{
Title
}
