% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{cfg}
\alias{cfg}
\title{Title}
\usage{
cfg(defineIntervals, variablesLong, invalidationCriteria, intervalPalette)
}
\arguments{
\item{defineIntervals}{}

\item{variablesLong}{}

\item{invalidationCriteria}{}

\item{intervalPalette}{}
}
\description{
Title
}
