% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{computeInvalids}
\alias{computeInvalids}
\title{Title}
\usage{
computeInvalids(Agregados1, invalidationCriteria)
}
\arguments{
\item{Agregados1}{}

\item{invalidationCriteria}{}
}
\description{
Title
}
