% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectInterval.R
\name{connectInterval}
\alias{connectInterval}
\title{Connect Clustered Time Intervals}
\usage{
connectInterval(interval, distance = dminutes(0), distanceRel = NULL)
}
\arguments{
\item{interval}{A data frame containing time intervals, usually with \code{start} and \code{end} POSIXct columns.}

\item{distance}{A \code{lubridate::Duration} or time span object. Intervals separated by a gap smaller than this absolute distance will be connected into a single interval. Default is 0 minutes.}

\item{distanceRel}{A numeric value. Connects intervals if the gap is smaller than a proportion (\code{distanceRel}) of the duration of the current or following interval. Default is NULL.}
}
\value{
A data frame with the connected/merged time intervals (\code{start} and \code{end}).
}
\description{
Merges consecutive or close time intervals into larger continuous bouts if the temporal gap between them is smaller than a specified absolute or relative threshold.
}
