% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteriaGenerator.R
\name{criteriaGenerator}
\alias{criteriaGenerator}
\title{Helps create functions for evaluating boolean expressions over a data frame that are going to be used as criteria for searching bouts}
\usage{
criteriaGenerator(EXPRESSION, useNW = TRUE)
}
\arguments{
\item{EXPRESSION}{}
}
\value{
A function that takes a data frame with accelerometry data and returns a vector of booleans
}
\description{
Helps create functions for evaluating boolean expressions over a data frame that are going to be used as criteria for searching bouts
}
