% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterio2Interval.R
\name{criterio2Interval}
\alias{criterio2Interval}
\title{Convert a data frame with a column called .criterio to a data frame of intervals of bouts of that criteria}
\usage{
criterio2Interval(df, pctBouts = 1, durBoutMin = dseconds(5),
  units = "mins")
}
\arguments{
\item{df}{data frame with columns .criterio}

\item{pctBouts}{}

\item{durBoutMin}{minimum amount of time that the condition muest be met to be aconsidered a Bout}

\item{units}{Units of time to show certain summaries. Une of c("secs","mins","hours","days")}
}
\value{
a list with data frame of intervals and certain summaries.
}
\description{
data frame of intervals of Bouts meeting certain criteria,
}
