% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterioBout.R
\name{criterioBout}
\alias{criterioBout}
\title{Indicates which rows meets the criteria for be considered part of a Bout meeting certain criteria}
\usage{
criterioBout(df, pctBouts = 1, durBoutMin = dseconds(5))
}
\arguments{
\item{df}{data frame with columns .criterio and eventually .criterioNW (that represents NonWear time as TRUE/FALSE).}

\item{pctBouts}{Numeric value between 0 and 1. Determines the minimum proportion or percentage of time within the window that the criteria must be fulfilled to be treated as a valid Bout.}

\item{durBoutMin}{Minimum amount of time (as a lubridate duration) that the condition must consistently be met to be considered a Bout.}
}
\value{
A boolean vector (TRUE/FALSE) indicating if the condition of belonging to a Bout is met for each corresponding epoch.
}
\description{
Generates a vector that indicates for each row of a dataframe (usually epoch or BIN file)
if that row meets the criteria to be considered part of a Bout (veryfy a criteria for ca percent of time of a minimum duration),
}
