% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterioZIB.R
\name{criterioZIB}
\alias{criterioZIB}
\title{Title}
\usage{
criterioZIB(df, critAnglez = 5, limSup = 30/1000, durBoutMin = dminutes(5))
}
\arguments{
\item{df}{}

\item{critAnglez}{}

\item{limSup}{}

\item{durBoutMin}{}
}
\description{
Title
}
