% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyDurationAdequate.R
\name{dailyDurationAdequate}
\alias{dailyDurationAdequate}
\title{Title}
\usage{
dailyDurationAdequate(whenList, whatList, durMin = NA, durMax = NA, ...)
}
\arguments{
\item{whenList}{}

\item{whatList}{}

\item{durMin}{}

\item{durMax}{}

\item{...}{}
}
\value{

}
\description{
Title
}
