% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyDurationWide.R
\name{dailyDurationWide}
\alias{dailyDurationWide}
\title{Title}
\usage{
dailyDurationWide(whenList, whatList, ...)
}
\arguments{
\item{whenList}{}

\item{whatList}{}

\item{...}{}
}
\description{
Title
}
