% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyQuantile.R
\name{dailyQuantile}
\alias{dailyQuantile}
\title{Title}
\usage{
dailyQuantile(whenList, whatList, q = 0.5, ...)
}
\arguments{
\item{whenList}{}

\item{whatList}{}

\item{q}{}

\item{...}{}
}
\value{

}
\description{
Title
}
