% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2list.R
\name{df2list}
\alias{df2list}
\title{convert a dataframe to a list of dataframes}
\usage{
df2list(df, label)
}
\arguments{
\item{df}{}

\item{label}{}
}
\value{

}
\description{
convert a dataframe to a list of dataframes
}
