% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{dygraphTS}
\alias{dygraphTS}
\title{Title}
\usage{
dygraphTS(
  df,
  intervals,
  UI_ejeY = "ENMO",
  runmeanType = NULL,
  runmeanTime = 0,
  definePalette = NULL,
  ...
)
}
\arguments{
\item{df}{}

\item{intervals}{}

\item{UI_ejeY}{}

\item{runmeanType}{}

\item{runmeanTime}{}

\item{definePalette}{}

\item{...}{}
}
\description{
Title
}
