% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmoBetween.R
\name{enmoBetween}
\alias{enmoBetween}
\title{Title}
\usage{
enmoBetween(
  df,
  limInf = 40/1000,
  limSup = 100/1000,
  pctBouts = 0.8,
  durBoutMin = dminutes(1)
)
}
\arguments{
\item{df}{}

\item{limInf}{}

\item{limSup}{}

\item{pctBouts}{}

\item{durBoutMin}{}
}
\value{

}
\description{
Title
}
