% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmoOver.R
\name{enmoOver}
\alias{enmoOver}
\title{Title}
\usage{
enmoOver(df, limInf = 100/1000, pctBouts = 0.8, durBoutMin = dminutes(1))
}
\arguments{
\item{df}{}

\item{limInf}{}

\item{pctBouts}{}

\item{durBoutMin}{}
}
\value{

}
\description{
Title
}
