% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmoUnder.R
\name{enmoUnder}
\alias{enmoUnder}
\title{Title}
\usage{
enmoUnder(df, limSup = 40/1000, pctBouts = 1, durBoutMin = dminutes(1))
}
\arguments{
\item{df}{}

\item{limSup}{}

\item{pctBouts}{}

\item{durBoutMin}{}
}
\value{

}
\description{
Title
}
