% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fechaHora2datetime.R
\name{fechaHora2datetime}
\alias{fechaHora2datetime}
\title{Title}
\usage{
fechaHora2datetime(calendardate, hora)
}
\arguments{
\item{calendardate}{}

\item{hora}{}
}
\value{

}
\description{
Title
}
