% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalDailySummary.R
\name{generalDailySummary}
\alias{generalDailySummary}
\title{Title}
\usage{
generalDailySummary(dfWhatWhen, variablesWhatWhen, functionDailySummary)
}
\arguments{
\item{dfWhatWhen}{}

\item{variablesWhatWhen}{}

\item{functionDailySummary}{}
}
\value{

}
\description{
Title
}
