% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genWhatL2_AinterruptedByB.R
\name{genWhatL2_AinterruptedByB}
\alias{genWhatL2_AinterruptedByB}
\title{Title}
\usage{
genWhatL2_AinterruptedByB(
  interval,
  A,
  B,
  pctBouts = 1,
  durBoutMin = dseconds(5)
)
}
\arguments{
\item{interval}{}

\item{A}{}

\item{B}{}

\item{pctBouts}{}

\item{durBoutMin}{}
}
\value{

}
\description{
Title
}
