% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genWhen_bedGGIR.R
\name{genWhen_bedGGIR}
\alias{genWhen_bedGGIR}
\title{Title}
\usage{
genWhen_bedGGIR(intervals, label = "bed")
}
\arguments{
\item{intervals}{}

\item{label}{}
}
\value{

}
\description{
Title
}
