% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalendars.R
\name{genWhen_bedJB}
\alias{genWhen_bedJB}
\title{Title}
\usage{
genWhen_bedJB(interval, label = "bed", durMin = dminutes(160),
  durMax = dhours(24))
}
\arguments{
\item{interval}{}

\item{label}{}

\item{durMin}{}

\item{durMax}{}
}
\description{
Title
}
