% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalendars.R
\name{genWhen_complementary}
\alias{genWhen_complementary}
\title{Title}
\usage{
genWhen_complementary(interval, When, label = "complementary")
}
\arguments{
\item{interval}{}

\item{When}{}

\item{label}{}
}
\description{
Title
}
