% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalendars.R
\name{genWhen_complementaryInternal}
\alias{genWhen_complementaryInternal}
\title{Title}
\usage{
genWhen_complementaryInternal(interval, When,
  label = "complementaryInternal")
}
\arguments{
\item{interval}{}

\item{When}{}

\item{label}{}
}
\description{
Title
}
