% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalendars.R
\name{genWhen_HorarioToHorarioFrom}
\alias{genWhen_HorarioToHorarioFrom}
\title{Title}
\usage{
genWhen_HorarioToHorarioFrom(horario1, horario2, label = "horario_1_2")
}
\arguments{
\item{horario1}{}

\item{horario2}{}

\item{label}{}
}
\description{
Title
}
