% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalendars.R
\name{genWhen_mostQuiet}
\alias{genWhen_mostQuiet}
\title{Title}
\usage{
genWhen_mostQuiet(interval, label = "mostQuiet")
}
\arguments{
\item{interval}{}

\item{label}{}
}
\value{

}
\description{
Title
}
