% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalendars.R
\name{genWhen_start2end}
\alias{genWhen_start2end}
\title{Title}
\usage{
genWhen_start2end(interval, label = "start2end")
}
\arguments{
\item{interval}{}

\item{label}{}
}
\value{

}
\description{
Title
}
