% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGGIR.R
\name{getGGIR}
\alias{getGGIR}
\title{Title}
\usage{
getGGIR(base, RAW, start = NA, end = NA, ...)
}
\arguments{
\item{base}{}

\item{RAW}{}

\item{start}{}

\item{end}{}

\item{...}{}
}
\description{
Title
}
