% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir2IntervalV3.R
\name{ggir2IntervalV3}
\alias{ggir2IntervalV3}
\title{Title}
\usage{
ggir2IntervalV3(currentProject, base, RAW, start = NA, end = NA)
}
\arguments{
\item{currentProject}{}

\item{base}{}

\item{RAW}{}

\item{start}{}

\item{end}{}
}
\description{
Title
}
