% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir2ts.R
\name{ggir2ts}
\alias{ggir2ts}
\title{Title}
\usage{
ggir2ts(base, RAW, start = NA, end = NA, progreso = NULL)
}
\arguments{
\item{base}{}

\item{RAW}{}

\item{start}{}

\item{end}{}

\item{progreso}{}
}
\description{
Title
}
