% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{girafePlot}
\alias{girafePlot}
\title{Title}
\usage{
girafePlot(
  intervals,
  WhenName = NA_character_,
  intervalToPlot = NULL,
  definePalette = NULL,
  offLimits = dseconds(0)
)
}
\arguments{
\item{intervals}{}

\item{WhenName}{}

\item{intervalToPlot}{}

\item{definePalette}{}

\item{offLimits}{}
}
\description{
Title
}
