% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{INA_Hildebrand_Child}
\alias{INA_Hildebrand_Child}
\title{Time in inactivity for Children (Hildebrand)}
\usage{
INA_Hildebrand_Child(df, ...)
}
\arguments{
\item{df}{Dataframe of epochs with two columns: datetime and ENMO}
}
\value{
dataframe of time intervals  matching the criteria
}
\description{
Find intervals of inactivity (INA, <1.5METs) such lying, sitting or standing.
Criteria: 
\itemize{
\item INA inactivity cut-off for children with ENMO metrics <36mg, according to Hildebrand 2017: Hildebrand M, Hansen BH, Van Hees VT, Ekelund U. Evaluation of raw acceleration sedentary thresholds in children and adults. Scand J Med Sci Sports. 2017 Dec;27(12):1814-1823. doi: 10.1111/sms.12795.
\item Non-dominant wrist cut-off point with Actigraph accelerometer.
\item 1-s epochs with 100\% of the time in INA.
}
}
