% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{interval_FromTo}
\alias{interval_FromTo}
\title{Title}
\usage{
interval_FromTo(
  intervalFrom,
  intervalTo,
  extremeOfFrom = "end",
  extremeOfTo = "start"
)
}
\arguments{
\item{intervalFrom}{}

\item{intervalTo}{}

\item{extremeOfFrom}{}

\item{extremeOfTo}{}
}
\description{
Title
}
