% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{intervalColor2checkbox}
\alias{intervalColor2checkbox}
\title{Title}
\usage{
intervalColor2checkbox(intervalColor, nombre, intervalSelected)
}
\arguments{
\item{intervalColor}{}

\item{nombre}{}

\item{intervalSelected}{}
}
\description{
Title
}
