% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{intervalQuantile2}
\alias{intervalQuantile2}
\title{Title}
\usage{
intervalQuantile2(dfIntervalos, probs = c(0, 0.25, 0.5, 0.75, 1))
}
\arguments{
\item{dfIntervalos}{}

\item{probs}{}
}
\description{
Title
}
