% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalsQuantiles.R
\name{intervalsQuantiles}
\alias{intervalsQuantiles}
\title{Calculate Temporal Quantiles of Intervals}
\usage{
intervalsQuantiles(intervals, probs = c(0.5))
}
\arguments{
\item{intervals}{A data frame representing the intervals (with \code{from} and \code{to} columns).}

\item{probs}{A numeric vector of probabilities (quantiles) with values between 0 and 1. Default is 0.5 (median).}
}
\value{
A data frame containing the corresponding timestamp quantiles for the overall duration of the provided intervals.
}
\description{
Computes cumulative temporal quantiles across a collection of intervals. It determines the exact timestamp where the specified duration proportions (\code{probs}) of the total interval time are reached.
}
