% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalidBedOfFirstDay.R
\name{invalidBedOfFirstDay}
\alias{invalidBedOfFirstDay}
\title{Title}
\usage{
invalidBedOfFirstDay(summary, whatCond = "isOn", whenInvalid = "bed",
  reason = "1stDayBed")
}
\arguments{
\item{summary}{}

\item{whatCond}{}

\item{whenInvalid}{}

\item{reason}{}
}
\description{
Title
}
