% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalidDuration.R
\name{invalidDuration}
\alias{invalidDuration}
\title{Title}
\usage{
invalidDuration(summary, whatCond, whenCond, minDur = NA, maxDur = NA,
  whenInvalid = NA_character_, reason = "unexplained")
}
\arguments{
\item{summary}{}

\item{whatCond}{}

\item{whenCond}{}

\item{minDur}{}

\item{maxDur}{}

\item{whenInvalid}{}

\item{reason}{}
}
\description{
Title
}
