% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{measureInterval_bed}
\alias{measureInterval_bed}
\title{measureInterval_bed es una definicion de cama alternativa a las que ya existen para considerar en el estudio de comparaciones}
\usage{
measureInterval_bed(df, SIB, parametersForBed = NULL, interruption = NULL, ...)
}
\arguments{
\item{df}{}

\item{SIB}{}

\item{parametersForBed}{}

\item{interruption}{}

\item{...}{}
}
\description{
measureInterval_bed es una definicion de cama alternativa a las que ya existen para considerar en el estudio de comparaciones
}
