% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{measureInterval_complementaryInternal}
\alias{measureInterval_complementaryInternal}
\title{measureInterval_complementaryInternal mide los intervalos complementarios a los que se les pase. Por ejemplo el complementario de estar dormido es estar despierto.}
\usage{
measureInterval_complementaryInternal(df, interval, ...)
}
\arguments{
\item{df}{}

\item{interval}{}

\item{...}{}
}
\description{
measureInterval_complementaryInternal mide los intervalos complementarios a los que se les pase. Por ejemplo el complementario de estar dormido es estar despierto.
}
