% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{MVPA_Hildebrand_Child}
\alias{MVPA_Hildebrand_Child}
\title{Time in Moderate and Vigorous Physical Activity for Children (Hildebrand)}
\usage{
MVPA_Hildebrand_Child(df, ...)
}
\arguments{
\item{df}{Dataframe of epochs with two columns: datetime and ENMO}
}
\value{
dataframe of time intervals matching the criteria
}
\description{
Find intervals of Moderate and Vigorous Physical Activity (MVPA, >3METs) such as brisk walking.
Criteria:   
\itemize{
\item MVPA cut-off for children with ENMO metrics, >200mg, according to Hildebrand 2014: Hildebrand M, Van Hees VT, Hansen BH, Ekelund U. Age-group comparibility of raw accelerometer output from wrist- and hipworn monitors. Med Sci Sports Exerc. 2014;46(9):1816–24.
\item Non-dominant wrist cut-off point with GENEActiv or Actigraph accelerometer.
\item 1-s epochs with 100\% of the time in MVPA.
}
}
