% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{palette2Color}
\alias{palette2Color}
\title{Title}
\usage{
palette2Color(listaColorVariable, intervalNames, ...)
}
\arguments{
\item{listaColorVariable}{}

\item{intervalNames}{}

\item{...}{}
}
\description{
Title
}
