% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOfIntervals.R
\name{plotOfIntervals}
\alias{plotOfIntervals}
\title{Title}
\usage{
plotOfIntervals(
  df,
  whatColumn = "what",
  whenColumn = "when",
  whenConcept = "daily",
  what2Color,
  epochColumnColor,
  ...
)
}
\arguments{
\item{df}{}

\item{whatColumn}{}

\item{whenColumn}{}

\item{whenConcept}{}

\item{what2Color}{}

\item{epochColumnColor}{}

\item{...}{}
}
\value{

}
\description{
Title
}
