% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantileForIntervals.R
\name{quantileForIntervals}
\alias{quantileForIntervals}
\title{Title}
\usage{
quantileForIntervals(df, q = 0.5, ...)
}
\arguments{
\item{df}{}

\item{q}{}

\item{...}{}
}
\value{

}
\description{
Title
}
