% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{readAccelerometryForPorto}
\alias{readAccelerometryForPorto}
\title{Title}
\usage{
readAccelerometryForPorto(
  path_ggir_ts = "",
  base = "",
  RAW = "",
  start = NA,
  end = NA,
  tz = "Europe/Madrid",
  ...
)
}
\arguments{
\item{path_ggir_ts}{}

\item{base}{}

\item{RAW}{}

\item{start}{}

\item{end}{}

\item{tz}{}

\item{...}{}
}
\description{
Title
}
