% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAccelerometryFromAccelerometer.R
\name{readAccelerometryFromAcclerometer}
\alias{readAccelerometryFromAcclerometer}
\title{Title}
\usage{
readAccelerometryFromAcclerometer(
  path_accelerometer_ts = "",
  baseBB = "",
  RAW = "",
  start = NA,
  end = NA,
  tz = "Europe/Madrid",
  ...
)
}
\arguments{
\item{RAW}{}

\item{start}{}

\item{end}{}

\item{tz}{}

\item{...}{}

\item{path_stepcount_ts}{}

\item{baseStepCount}{}
}
\description{
Title
}
