% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{readAccelerometryFromBioBank}
\alias{readAccelerometryFromBioBank}
\title{Title}
\usage{
readAccelerometryFromBioBank(
  path_biobank_ts = "",
  baseBB = "",
  RAW = "",
  start = NA,
  end = NA,
  tz = "Europe/Madrid",
  ...
)
}
\arguments{
\item{path_biobank_ts}{}

\item{baseBB}{}

\item{RAW}{}

\item{start}{}

\item{end}{}

\item{tz}{}

\item{...}{}
}
\description{
Title
}
