% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{readAccelerometryFromStepCount}
\alias{readAccelerometryFromStepCount}
\title{Title}
\usage{
readAccelerometryFromStepCount(
  path_stepcount_ts = "",
  baseStepCount = "",
  RAW = "",
  start = NA,
  end = NA,
  tz = "Europe/Madrid",
  ...
)
}
\arguments{
\item{path_stepcount_ts}{}

\item{baseStepCount}{}

\item{RAW}{}

\item{start}{}

\item{end}{}

\item{tz}{}

\item{...}{}
}
\description{
Title
}
