% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{searchMaxActivity}
\alias{searchMaxActivity}
\title{Title searchMaxActivity busca dentro de un intervalo de medida el tramo en que ocurre la mayor cantidad de cierta variable de interés}
\usage{
searchMaxActivity(
  df,
  intervals,
  durInterval = dminutes(60),
  colActivity = "step_count"
)
}
\arguments{
\item{df}{}

\item{intervals}{}

\item{durInterval}{}

\item{colActivity}{}
}
\description{
Title searchMaxActivity busca dentro de un intervalo de medida el tramo en que ocurre la mayor cantidad de cierta variable de interés
}
