% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseWhatByWhen.R
\name{summariseWhatByWhen}
\alias{summariseWhatByWhen}
\title{Title}
\usage{
summariseWhatByWhen(When, What, by = "day")
}
\arguments{
\item{When}{}

\item{What}{}

\item{by}{}
}
\value{

}
\description{
Title
}
