% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts2IntervalV2.R
\name{ts2IntervalV2}
\alias{ts2IntervalV2}
\title{Title}
\usage{
ts2IntervalV2(ts, currentProject, nucleos = numCores(4), ...)
}
\arguments{
\item{ts}{}

\item{currentProject}{}

\item{nucleos}{}

\item{...}{}
}
\value{

}
\description{
Title
}
