% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{verificar_cambio_horario}
\alias{verificar_cambio_horario}
\title{Title}
\usage{
verificar_cambio_horario(fecha_inicio, fecha_fin)
}
\arguments{
\item{fecha_inicio}{}

\item{fecha_fin}{}
}
\description{
Title
}
