% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{VPA_Hildebrand_Child}
\alias{VPA_Hildebrand_Child}
\title{Time in Vigorous Physical Activity for Children (Hildebrand)}
\usage{
VPA_Hildebrand_Child(df)
}
\arguments{
\item{df}{Dataframe of epochs with two columns: datetime and ENMO}
}
\value{
dataframe of time interval matching the criteria
}
\description{
Find intervals of Vigorous Physical Activity (VPA, >6METs) such as jogging or running.
Criteria: 
\itemize{
\item VPA cut-off for children with ENMO metrics, >700mg, according to Hildebrand 2014: Hildebrand M, Van Hees VT, Hansen BH, Ekelund U. Age-group comparibility of raw accelerometer output from wrist- and hipworn monitors. Med Sci Sports Exerc. 2014;46(9):1816–24.
 \item Non-dominant wrist cut-off point with GENEActiv or Actigraph accelerometer.
 \item  1-s epochs with 100\% of the time in VPA.
}
}
