% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatAlgebra.R
\name{whatDifference}
\alias{whatDifference}
\title{Title}
\usage{
whatDifference(intervalSpace, intervalRemove)
}
\arguments{
\item{intervalSpace}{}

\item{intervalRemove}{}
}
\value{

}
\description{
Title
}
