% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicVariables.R
\name{ZIB}
\alias{ZIB}
\title{Title}
\usage{
ZIB(df, critAnglez = 5, durBoutMin = dminutes(5), ...)
}
\arguments{
\item{df}{}

\item{critAnglez}{}

\item{durBoutMin}{}

\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
 
}
