% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerometry2Intervals.R
\name{accelerometry2Intervals}
\alias{accelerometry2Intervals}
\title{accelerometry2Intervals}
\usage{
accelerometry2Intervals(driver, defineIntervals, activity_log, ...)
}
\description{
accelerometry2Intervals
}
