% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{aggregate2_Duration_meansd}
\alias{aggregate2_Duration_meansd}
\title{Title}
\usage{
aggregate2_Duration_meansd(df, ...)
}
\arguments{
\item{df}{}

\item{...}{}
}
\value{
mean_dur : media de las duraciones, sd_dur: desviación estándar de las duraciones
}
\description{
Title
}
