% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{aggregate_Duration}
\alias{aggregate_Duration}
\title{Title}
\usage{
aggregate_Duration(df, units = "secs", ...)
}
\arguments{
\item{df}{}

\item{units}{}

\item{...}{}
}
\value{
Devuelve las columna: name, dur
}
\description{
Title
}
