% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectInterval.R
\name{connectInterval}
\alias{connectInterval}
\title{Title}
\usage{
connectInterval(interval, distance = dminutes(0), distanceRel = NULL)
}
\arguments{
\item{interval}{}

\item{distance}{}

\item{distanceRel}{}
}
\description{
Title
}
