% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfoOfFiles.R
\name{getInfoOfFiles}
\alias{getInfoOfFiles}
\title{Title}
\usage{
getInfoOfFiles(dir, prefix = "", suffix = "\\\\.RData$")
}
\arguments{
\item{dir}{}

\item{prefix}{}

\item{suffix}{}
}
\value{

}
\description{
Title
}
