% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterio2Summary.R, R/getSummary.R
\name{getSummary}
\alias{getSummary}
\title{Create a summery for the column .criterio of a data frame representing accelerometer data}
\usage{
getSummary(
  df,
  offset = dhours(0),
  minimoHorasValidas = 20,
  maximoHorasNonWear = 2,
  durBoutMin = dseconds(5)
)

getSummary(
  df,
  offset = dhours(0),
  minimoHorasValidas = 20,
  maximoHorasNonWear = 2,
  durBoutMin = dseconds(5)
)
}
\arguments{
\item{df}{}

\item{offset}{}

\item{minimoHorasValidas}{}

\item{maximoHorasNonWear}{}

\item{durBoutMin}{}
}
\value{
a list with a summary. This is the info that we use to define activity variables on a daily basis and on a global value.
}
\description{
Generates a summary for a epoch or BIN file
}
