% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectIntervals.R
\name{intersectIntervals}
\alias{intersectIntervals}
\title{Title}
\usage{
intersectIntervals(what_n, when_n, useNames = FALSE, short = TRUE)
}
\arguments{
\item{what_n}{}

\item{when_n}{}

\item{useNames}{}

\item{short}{}
}
\description{
Title
}
