% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalsQuantiles.R
\name{intervalsQuantiles}
\alias{intervalsQuantiles}
\title{Title}
\usage{
intervalsQuantiles(intervals, probs = c(0.5))
}
\arguments{
\item{intervals}{}

\item{probs}{}
}
\value{

}
\description{
Title
}
