% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedInterval_01h.R
\name{measureInterval_01h}
\alias{measureInterval_01h}
\title{measureInterval_01h}
\usage{
measureInterval_01h(
  df,
  first = TRUE,
  last = TRUE,
  offsetLabels = dhours(0),
  starts = dhours(0),
  duration = period(1, "hours")
)
}
\description{
measureInterval_01h
}
