% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedInterval_24h.R
\name{measureInterval_24h}
\alias{measureInterval_24h}
\title{measureInterval_24h}
\usage{
measureInterval_24h(
  df,
  first = TRUE,
  last = TRUE,
  offsetLabels = dhours(0),
  starts = dhours(0),
  duration = period(1, "days"),
  .isOn = NULL,
  withName = TRUE
)
}
\description{
measureInterval_24h
}
