% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAccelerometryFromGGIR.R
\name{readAccelerometryFromGGIR}
\alias{readAccelerometryFromGGIR}
\title{Title}
\usage{
readAccelerometryFromGGIR(
  path_ggir_ts = "",
  base = "",
  RAW = "",
  start = NA,
  end = NA,
  tz = "Europe/Madrid",
  ...
)
}
\arguments{
\item{base}{}

\item{RAW}{}

\item{start}{}

\item{end}{}

\item{tz}{}

\item{...}{}
}
\description{
Title
}
