% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchByCriteria.R
\name{searchByCriteria}
\alias{searchByCriteria}
\title{Title: searchByCriteria}
\usage{
searchByCriteria(
  df,
  EXPRESSION,
  pctBouts = 0.8,
  durBoutMin = dminutes(1),
  runmean = FALSE,
  useNW = TRUE
)
}
\arguments{
\item{df}{}

\item{EXPRESSION}{}

\item{pctBouts}{}

\item{durBoutMin}{}
}
\description{
Title: searchByCriteria
}
