% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFunctions.R
\name{SIB_v2}
\alias{SIB_v2}
\title{Title}
\usage{
SIB_v2(df, critAnglez = 5, durBoutMin = dminutes(5), ...)
}
\arguments{
\item{df}{}

\item{critAnglez}{}

\item{durBoutMin}{}

\item{...}{}
}
\description{
Title
}
