% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectIntervals.R
\name{touchedIntervals}
\alias{touchedIntervals}
\title{Title touchedIntervals}
\usage{
touchedIntervals(what_n, when_n, useNames = FALSE, short = TRUE, operator = I)
}
\arguments{
\item{what_n}{}

\item{when_n}{}

\item{useNames}{}

\item{short}{}

\item{operator}{}
}
\description{
Title touchedIntervals
}
